Allow searching for the exact value
All search terms have an equivalent suffixed with `__exact` which will
search for the exact string. The only exception is the `email` search
term which always searches for the exact string.

The `creation_before` search term has been renamed to
`creation__before` for coherence. This is technically a backwards
incompatible change, but the term has only been added a few days ago and
not advertised, so I'm fairly confident noone uses it yet.

Fixes: #266

Signed-off-by: Aurélien Bompard <aurelien@bompard.org>